# api/docs/swagger.yaml
...
info:
  contact: {}
  description: API for real-time messaging via WebSocket
  license:
    name: GNU 3.0
    url: https://github.com/IdanKoblik/whisper/blob/main/LICENSE.md
  title: Whisper API
  version: 2025-b2
paths:
  /admin/register:
    post:
      consumes:
      - application/json
      description: Create a new API token for authentication. Requires admin privileges.
        Returns the generated API token.
      produces:
      - text/plain
      responses:
        "201":
          description: API token created successfully
          schema:
            type: string
        "400":
          description: Failed to create token
          schema:
            type: string
        "401":
          description: Unauthorized - Admin access required
          schema:
            type: string
      security:
      - AdminAuth: []
      summary: Register new API token
      tags:
      - admin
...
securityDefinitions:
  AdminAuth:
    description: Admin token for administrative operations
    in: header
    name: X-Admin-Token
    type: apiKey
  ApiKeyAuth:
    description: API token for authentication
    in: header
    name: X-API-Token
    type: apiKey
swagger: "2.0"